Malware Detector Nightmare Project
A Python project that had to use different classifiers for detecting malware. Categorizes the test class as True Positive, True Negative, False Positive, and False Negative. I learned python by throwing myself into the fire on this project and completing it was truly a nightmare. This was a group project with another student.